requirements.txt
fastapi
uvicorn
pydantic
openenv-core
openai
streamlit
plotly
pandas
stable-baselines3
torch
gymnasium
psutil
docker
sqlalchemy
python-dotenv
numpy
scipy
requests
psycopg2-binary
